pyproject.toml
[tool.poetry]
name = "deep-lab3"
version = "0.1.0"
description = ""
authors = ["ArnaudGansmandel <a.gansmandel@laposte.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ipykernel = "^6.29.5"
scikit-learn = "^1.5.1"
keras = "^3.4.1"
requests = "^2.32.3"
pillow = "^10.4.0"
matplotlib = "^3.9.1"
pyyaml = "^6.0.1"
h5py = "^3.11.0"
tensorflow-intel = "^2.17.0"
pandas = "^2.2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
